refactor AddNewWorkPlanModal into AddEditReportModal
