+++ services/ui-src/src/components/pages/Dashboard/DashboardPage.tsx
@@ -16,7 +16,6 @@ import {
   useDisclosure,
 } from "@chakra-ui/react";
 import {
-  AddNewWorkPlanModal,
   AddEditReportModal,
   Modal,
   DashboardTable,
@@ -206,13 +205,7 @@ export const DashboardPage = ({ reportType }: Props) => {
     }
 
     setSelectedReport(formData);
-
-    // use disclosure to open modal
-    if (!reportsToDisplay?.length && reportType == ReportType.WP) {
-      addNewWorkPlanModalOnOpenHandler();
-    } else {
-      addEditReportModalOnOpenHandler();
-    }
+    addEditReportModalOnOpenHandler();
   };
 
   const toggleReportArchiveStatus = async (report: ReportShape) => {
@@ -263,13 +256,6 @@ export const DashboardPage = ({ reportType }: Props) => {
         return true;
     }
   };
-
-  // new work plan modal disclosure
-  const {
-    isOpen: addNewWorkPlanModalIsOpen,
-    onOpen: addNewWorkPlanModalOnOpenHandler,
-    onClose: addNewWorkPlanModalOnCloseHandler,
-  } = useDisclosure();
 
   // add/edit program modal disclosure
   const {
@@ -383,15 +369,6 @@ export const DashboardPage = ({ reportType }: Props) => {
           </Box>
         )}
       </Box>
-      <AddNewWorkPlanModal
-        activeState={activeState!}
-        selectedReport={selectedReport!}
-        reportType={reportType}
-        modalDisclosure={{
-          isOpen: addNewWorkPlanModalIsOpen,
-          onClose: addNewWorkPlanModalOnCloseHandler,
-        }}
-      />
       <AddEditReportModal
         activeState={activeState!}
         selectedReport={selectedReport!}
